+++ xLights/outputs/xxxEthernetOutput.h
@@ -61,7 +61,7 @@ public:
 
     virtual int32_t GetMaxChannels() const override { return xxxETHERNET_MAX_CHANNELS; }
     static int GetMaxxxxChannels() { return xxxETHERNET_MAX_CHANNELS; }
-    virtual bool IsValidChannelCount(int32_t channelCount) const override { return channelCount > 0 && channelCount <= xxxETHERNET_MAX_CHANNELS; }
+    virtual bool IsValidChannelCount(int32_t channelCount) const override { return channelCount > 0 && static_cast<uint32_t>(channelCount) <= xxxETHERNET_MAX_CHANNELS; }
 
     int GetId() const { return _universe; }
     void SetId(int id) { _universe = id; _dirty = true; }
